Move ApiError to service package
